支払い状況をグラフ化した
