+++ native/android/app/src/main/res/layout/fragment_add_payment_step2.xml
@@ -11,7 +11,7 @@
       android:id="@+id/tab_layout"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
-      android:background="@color/ourTheme"
+      android:background="@color/completedPayment"
       app:layout_constraintBottom_toTopOf="@+id/viewPager2"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
